requirements/base.txt
# Dependencias base del proyecto (core dependencies)

# === Serialización y Formato de Datos ===
# PyYAML: Parser y emisor de YAML para archivos de configuración
PyYAML>=6.0.0

# === GUI Moderna ===
# customtkinter: Framework moderno para interfaces gráficas
customtkinter>=5.2.0

# === Utilidades de Sistema ===
# click: Creación de interfaces de línea de comandos
click>=8.0.0
